+++ src/parser/locale/ja.test.ts
@@ -1,0 +1,15 @@
+﻿import { japanese } from './ja';
+
+describe('Unit: parser/locale/ja', () => {
+  describe('japanese', () => {
+    it('ja', () => {
+      assert(japanese('、'));
+      assert(japanese('。'));
+      assert(japanese('！'));
+      assert(japanese('？'));
+      assert(japanese('…'));
+    });
+
+  });
+
+});