+++ src/parser/locale/ja.ts
@@ -1,0 +1,6 @@
+﻿const endings = /[、。！？…]/;
+
+export function japanese(char: string): boolean {
+  assert([...char].length == 1);
+  return char.search(endings) === 0;
+}