+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.58.0
 
+- Localize linebreak processing.
 - Remove linebreaks after `wbr` html tag.
 
 ## 0.57.0